Clarifies scale and offset use in the LFO modules.
I have realised that the OFFSET module first sums the offset and then multiplies by the scale, as is written clearly in the OFFSET documentation.

But for the LLFO module, scale is a scale _percentage_ and, it is applied _before_ the offset, not after. I presume that the other LFO do the same, but it's not written in the documentation. Perhaps it is obvious, and it is a problem for me just because I'm a newbie? 

Thanks for all the amazing work!